Music/jobboard/Django-project/requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.2.3
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
package-name==0.1
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
